Group exec and inspect functionality.
